Rename assets directory
