+++ app/src/App.js
@@ -26,7 +26,7 @@ import Status from './pages/Status'
 import Footer from './components/Footer'
 import FeedbackModal from './components/FeedbackModal.js'
 import GDPRModal from './components/GDPRModal.js'
-import AssetsPage from './pages/assets/AssetsPage'
+import AssetsPage from './pages/Assets/AssetsPage'
 import { getConsentUrl } from './utils/utils'
 import './App.scss'
