+++ app/src/pages/Assets/Assets.js
