+++ app/src/pages/Assets/AssetsPage.js
