+++ app/src/pages/Assets/AssetsPageNavigation.css
