+++ app/src/pages/Assets/AssetsTable.js
