+++ app/src/pages/Assets/AssetsTableErrorMessage.js
